requirements.txt
flask==2.3.3
python-docx==0.8.11
markdown==3.4.4
gunicorn==21.2.0
Werkzeug==2.3.7
PyPDF2==3.0.1